requirements.txt
python-telegram-bot~=21.0
PyQt5~=5.15
openai~=1.0
python-dotenv~=1.0